build-kernel.sh: Handle kernel configs with CONFIG_LOCALVERSION
Setting this option in your kernel config will result in filenames that
build-kernel.sh does not expect. To not introduce confusion between
recipe versions and debian package versions simply require the recipe to
have a PV matching "make kernelrelease".

+++ meta/recipes-kernel/linux/files/build-kernel.sh
@@ -21,6 +21,13 @@ flock /dpkg.lock \
 cd $1
 make olddefconfig
 
+KV=$( make -s kernelrelease )
+if [ "${KV}" != "${PV}" ]; then
+	echo "ERROR: Recipe PV is \"${PV}\" but should be \"${KV}\"" 1>&2
+	echo "ERROR: Probably due to CONFIG_LOCALVERSION" 1>&2
+	exit 1
+fi
+
 rm -f .version
 make -j $(($(nproc) * 2)) deb-pkg
